refactored to subdivide large buffers and manage pipelines and pipeline resources better

+++ src/Material.cpp
@@ -1,3 +1,3 @@
 #include "Material.h"
 
-std::unordered_map<size_t, Material> Material::materials;
+std::unordered_map<uint64_t, Material> Material::materials;
